some changes about folllow andt followrequest

+++ src/main/java/com/example/socialmedia/service/impl/FollowRequestServiceImpl.java
@@ -3,6 +3,7 @@ package com.example.socialmedia.service.impl;
 import com.example.socialmedia.model.FollowRequest;
 import com.example.socialmedia.model.User;
 import com.example.socialmedia.repository.FollowRequestRepository;
+import com.example.socialmedia.repository.UserRepository;
 import com.example.socialmedia.service.AuthService;
 import com.example.socialmedia.service.FollowRequestService;
 import lombok.RequiredArgsConstructor;
@@ -19,6 +20,7 @@ public class FollowRequestServiceImpl implements FollowRequestService {
 
     private final FollowRequestRepository followRequestRepository;
     private final AuthService authService;
+    private final UserRepository userRepository;
 
     @Override
     public FollowRequest getById(int id) {
@@ -36,8 +38,12 @@ public class FollowRequestServiceImpl implements FollowRequestService {
         if(followRequest==null){
             return new ResponseEntity<>("Not found Follow Request with given id", HttpStatus.NOT_FOUND);
         }
-        User user = authService.getById(followRequest.getTargetUserId());
-        currentUser.getFollowers().add(user);
+        if(currentUser.getId()!=followRequest.getTargetUserId()){
+            return new ResponseEntity<>("This follow request is not your",HttpStatus.FORBIDDEN);
+        }
+        User user = authService.getById(followRequest.getSourceUser().getId());
+        user.getFollowers().add(currentUser);
+        userRepository.save(user);
         followRequestRepository.delete(followRequest);
         return new ResponseEntity<>("Follow Request has been accepted",HttpStatus.OK);
     }
